+++ CHANGELOG.md
@@ -11,6 +11,8 @@ but cannot always guarantee backwards compatibility. Changes that may **break co
 
 **Improved**
 
+- 🚀 Added new forecasting model `T0Model` : [The Forecasting Company's open-weights ~100M-parameter foundation model](https://huggingface.co/theforecastingcompany/t0-alpha) for zero-shot forecasting. It supports univariate, multivariate, and multiple time series as well as past and future covariates, without training, and can output deterministic or probabilistic forecasts. It can also be fine-tuned (full or partial) with `enable_finetuning`. [#3142](https://github.com/unit8co/darts/pull/3142) by [Geoffrey Négiar](https://github.com/GeoffNN), [Huikan Xiang](https://github.com/huikan-tfc) and [Lucas Meyer](https://github.com/LTMeyer).
+
 **Fixed**
 
 **Dependencies**